Modified the charlieplexing code to utilise compacted arrays, generated by the new python script. Other minor changes, like replacing the use of uint16_t with uint8_t where possible. Binary size drastically reduced

+++ src/model/model.c
@@ -38,7 +38,7 @@ ISR(TIMER1_COMPA_vect) {
 
 
 void incrementDisplayMode() {
-  uint16_t currentMode = GET_VALUE(clicksAndFlags, DISPLAY_MODE);
+  uint8_t currentMode = GET_VALUE(clicksAndFlags, DISPLAY_MODE);
   currentMode = currentMode < (NUMBER_OF_MODES - 1) ? currentMode + 1 : 0;
   clicksAndFlags &= ~DISPLAY_MODE_MASK;
   clicksAndFlags |= (currentMode << DISPLAY_MODE_OFFSET);
